Fix wrapping on options with long help messages (#11)
Fixes #10

The problem was the wrap width was calculated based on the space remaining after the option names. This works for the first line, but subsequent lines have a width that's shorter than it should be.

e.g. 

```
              width=17
          /---------------\ 
--option  lorem ipsum dolor
          sit
          amet
\--------------/
    width=17
```

The fix is to use the option name and spacing as the initial indent, and wrap to the full width:

```
        width=28
/-------------------------\ 
  inital indent
/--------\
--option  lorem ipsum dolor
          sit amet
\-------------------------/
       width=28