Split Editor#editorPath and pass it as program and arguments to ProcessBuilder()
ProcessBuilder searches for an executable file in the PATH withe exactly the name provided.
For strings like "atom --wait" this means it would e.g. in /bin search for a
"/bin/atom --wait" binary which does not exist.

This commit provides ProcessBuilder the program as the first item in an array. Options like
"--wait" together with the filename are provided as further array items. The above example
 "atom --wait" becomes ProcessBuilder("atom", "--wait", filename).

Fixes #97