Add FlagOption.convert (#210)
The semantics of this are different from `OptionWithValues.convert`. The latter conversion is called one for each occurrence of the option on the command line, while this flag conversion uses `transformAll`, and so is only called once with the final value of the flag.

We could have added a `ValueTransformer` field to `FlagOption` to make the semantics match, but that wouldn't be backwards compatible. I don't think it makes much sense for flags to have a value transformer anyway; this design feels more intuitive.