Fix error message when lazily referencing required options (#431)
In order to support referencing other options in `defaultLazy`, we call `finalize` twice if the first call caused an IllegalStateException. The second time happens after all other options are finalized, so as long as there are no reference cycles or chains, the second call will succeed.

In order to support MultiUsageError, we have to collect all UsageErrors until after the second finalization round.

The regression happened because we were throwing IllegalStateExceptions that occur in the second round even if there were deferred UsageErrors.